Capitalize function now in correct place
