--- utils/strings.spec.ts
@@ -1,8 +1,0 @@
-import { capitalize } from './strings';
-
-describe('strings utils', () => {
-  it('should capitalize correctly words', () => {
-    expect(capitalize('this is a string')).toBe('This is a string');
-    expect(capitalize('fooobar')).toBe('Fooobar');
-  });
-});
