--- utils/strings.tsx
@@ -1,4 +1,0 @@
-export const capitalize = (str: string) => {
-  const lower = str.toLowerCase();
-  return str.charAt(0).toUpperCase() + lower.slice(1);
-};
